+++ data-migrations/2026-04-15-0949-add-provenance-to-projects.sql
@@ -1,0 +1,4 @@
+-- Adds a provenance column to projects
+ALTER TABLE projects
+  ADD COLUMN provenance VARCHAR(50) NOT NULL DEFAULT 'dmptool' AFTER id,
+  ADD INDEX projects_provenance_idx (provenance);